move parseLinkHeader to util.js, set ?per_page=100 and parse link headers to fetch all branches
